requirements.txt
jax>=0.4.23
jaxlib>=0.4.23
orbax-checkpoint>=0.5.5
absl-py
array-record
aqtp
blobfile
cloud-accelerator-diagnostics
cloud-tpu-diagnostics
google-cloud-storage
grain-nightly
flax>=0.8.0
ml-collections
ml-goodput-measurement
numpy
optax
protobuf==3.20.3
pylint
pytest
pyink
pre-commit
pytype
sentencepiece==0.1.97
tensorflow-text>=2.13.0
tensorflow>=2.13.0
tensorflow-datasets
tensorboardx
tensorboard-plugin-profile
tiktoken
git+https://github.com/mlperf/logging.git
google-jetstream
